transformers/torchvision_preprocess/requirements.txt
aistore[etl]>=1.14.0
pillow>=10.0.0,<10.1.0
torch>=2.0.0
torchvision>=0.15.0